// models/endpoints.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Bandwidth
 * Bandwidth\'s Communication APIs
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: letstalk@bandwidth.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { EndpointStatusEnum } from './endpoint-status-enum';
// May contain unused imports in some cases
// @ts-ignore
import type { EndpointTypeEnum } from './endpoint-type-enum';

export interface Endpoints {
    /**
     * The unique ID of the endpoint.
     */
    'endpointId': string;
    'type': EndpointTypeEnum;
    'status': EndpointStatusEnum;
    /**
     * The time the endpoint was created. In ISO-8601 format.
     */
    'creationTimestamp': string;
    /**
     * The time the endpoint token will expire. In ISO-8601 format. Tokens last 24 hours.
     */
    'expirationTimestamp': string;
    /**
     * A tag for the endpoint.
     */
    'tag'?: string;
}