// models/page.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Bandwidth
 * Bandwidth\'s Communication APIs
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: letstalk@bandwidth.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface Page {
    /**
     * The number of items per page.
     */
    'pageSize': number;
    /**
     * The total number of items.
     */
    'totalElements'?: number;
    /**
     * The total number of pages.
     */
    'totalPages'?: number;
    /**
     * The current page number.
     */
    'pageNumber'?: number;
}